missing in the last commit, use qgis.db instead of config file to store transformations

+++ transformationManagerDlg.py
@@ -1,3 +1,25 @@
+# -*- coding: utf-8 -*-
+
+"""
+/***************************************************************************
+Name                 : Transformation tools
+Description          : Help to use grids and towgs84 to transform a vector/raster
+Date                 : April 16, 2011 
+copyright            : (C) 2011 by Giuseppe Sucameli (Faunalia)
+email                : brush.tyler@gmail.com
+
+ ***************************************************************************/
+
+/***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ ***************************************************************************/
+"""
+
 from PyQt4.QtCore import *
 from PyQt4.QtGui import *
 
@@ -8,6 +30,7 @@ class TransformationManagerDlg(QDialog, Ui_Dialog):
 	
 	def __init__(self, iface, parent=None):
 		QDialog.__init__(self, parent)
+		self.setWindowFlags( Qt.Window | Qt.WindowTitleHint | Qt.WindowSystemMenuHint | Qt.WindowCloseButtonHint | Qt.WindowMaximizeButtonHint & ~Qt.WindowMinimizeButtonHint )
 		self.setAttribute(Qt.WA_DeleteOnClose)
 		self.iface = iface
 		self.setupUi(self)
@@ -48,6 +71,10 @@ class TransformationManagerDlg(QDialog, Ui_Dialog):
 		if not index.isValid():
 			return
 		t = self.table.model().getAtRow( index.row() )
+		res = QMessageBox.question(self, self.tr(u"Delete transformation"), self.tr(u"Really delete transformation '%1'?" ).arg(t.name), QMessageBox.Yes | QMessageBox.No)
+		if res == QMessageBox.No:
+			return
+
 		t.deleteData()
 		self.refreshTable()
